+++ src/form/RruMultiSelectInput/RruMultiSelectInput.tsx
@@ -31,6 +31,13 @@ const RruMultiSelectInput: FC<RruMultiSelectInputProps> = (props) => {
   const field = useField(props.name, (serializedValue) => {
     if (!deepEqual(serializedValue, value)) {
       setValue(serializedValue);
+
+      if (!deepEqual(serializedValue, selectedOptions.map(o => o.value))) {
+        setTimeout(() => {
+          const options = findOptions(props.options, serializedValue || []);
+          onSelectChange(options);
+        }, 0);
+      }
     }
   });